requirements.txt
h5py==3.8.0
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
requests==2.28.2
streamlit==1.12.2
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
tensorflow==2.11.0
